# back/api/v1/openapi.yaml
openapi: 3.0.3
...
info:
  title: cursed_matrix API
  version: 0.2.0
  description: |
    Every timestamp is ISO-8601 with an explicit UTC offset. Calendar days are
    YYYY-MM-DD strings computed in the user's timezone, never the server's.
    The server returns codes and parameters, never rendered prose: the client
    renders them in EN or RU.
servers:
  - url: /api/v1
...
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the token pair
      operationId: refresh
      description: |
        The refresh cookie is the credential; there is no body. A token that was
        already consumed is a replay: the server revokes every refresh token of
        that account and answers 401.
      security:
        - refreshCookie: []
      responses:
        "204":
          description: New cookies are set.
          headers:
            Set-Cookie: { $ref: "#/components/headers/SessionCookies" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /auth/logout-all:
    post:
      tags: [auth]
      summary: Sign out everywhere
      operationId: logoutAll
      description: |
        Revokes every refresh token of the account. Other devices keep working
        until their access token expires, which is what bounds the damage of a
        stolen cookie to fifteen minutes.
      responses:
        "204":
          description: Revoked.
          headers:
            Set-Cookie: { $ref: "#/components/headers/SessionCookies" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
...
  /me/export:
    post:
      tags: [account]
      summary: Take your data away
      operationId: exportAccount
      description: |
        Everything the account holds, in one document: the account itself, its
        settings and figures, every task including the deleted ones, the links,
        the labels and the achievements. Read in one transaction, so the pieces
        agree with each other.

        Synchronous, which the free plan's size makes reasonable. An account
        large enough to need a job would answer 202 with a link instead, and
        the two shapes are told apart by the status.
      responses:
        "200":
          description: The export.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountExport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
  /me/delete:
    post:
      tags: [account]
      summary: Close the account
      operationId: deleteAccount
      description: |
        Asks for the password again: `docs/API.md` calls for a token sent by
        email, but an address is optional in this product — the account is
        identified by its username — so most accounts have nowhere to send
        one. The password is something only the owner knows, which is what
        stops a borrowed session from ending an account.

        The delete is logical. Nothing earned is erased, because the XP ledger,
        the activity history and other people's leaderboard standings all refer
        to the row; what ends is access, and every ranking already excludes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "204": { description: Closed. }
        "401":
          description: No session, or the password does not match.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
...
  /me:
    get:
      tags: [account]
      summary: The signed-in account
      operationId: getMe
      description: |
        What the client boots from: it drives the header, the subscription gate
        and the profile. `level` is derived from `lifetimeXp` by the server; the
        client mirrors the formula for display only.
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    patch:
      tags: [account]
      summary: Change account preferences
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsUpdate" }
      responses:
        "200":
          description: The account, with the new preferences applied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CSRFRejected" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: cm_access
      description: |
        A signed JWT the page cannot read. Verified by signature, so an
        authenticated request touches neither PostgreSQL nor Redis.
    refreshCookie:
      type: apiKey
      in: cookie
      name: cm_refresh
      description: |
        The opaque id of a server-side record, scoped to /api/v1/auth and
        SameSite=Strict. Revocable, which is what the JWT cannot be.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        The value of the page-readable cm_csrf cookie. Required on every unsafe
        method, because a browser attaches cookies whether or not the page meant
        to send them.
  headers:
    SessionCookies:
      description: |
        Sets or clears cm_access (HttpOnly, SameSite=Lax, Path=/api),
        cm_refresh (HttpOnly, SameSite=Strict, Path=/api/v1/auth) and cm_csrf.
      schema:
        type: string
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    CSRFRejected:
      description: The CSRF header is missing or does not match the cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: Valid JSON, invalid content.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code]
          properties:
            code:
              type: string
              description: |
                Stable and language-independent. The client renders it; the
                server never sends prose.
              example: VALIDATION_FAILED
            params:
              type: object
              additionalProperties: true
              description: Values for the client's message template.
            requestId:
              type: string
              description: What the user reports when something goes wrong.
...
    SettingsUpdate:
      type: object
      description: Only the fields present are changed.
      properties:
        language: { $ref: "#/components/schemas/Language" }
        timezone: { type: string, example: Europe/Berlin }
        email:
          type: string
          format: email
          nullable: true
          description: |
            Adding, changing or clearing the address. Uniqueness applies only
            where there is one.
        showInLeaderboard:
          type: boolean
          description: |
            Taking part in the public ranking is voluntary. Turning it off
            removes the user from the ranking without touching their XP.
    Language:
      type: string
      enum: [EN, RU]
    Quadrant:
      type: string
      enum: [IMPORTANT_URGENT, IMPORTANT_NOT_URGENT, NOT_IMPORTANT_URGENT, NOT_IMPORTANT_NOT_URGENT]
    TaskColor:
      type: string
      enum: [NONE, CYAN, VIOLET, AMBER, ROSE, TEAL, SLATE]
    TaskStatus:
      type: string
      enum: [ACTIVE, COMPLETED]
    CompletionSource:
      type: string
      enum: [DIRECT, PARENT_CASCADE]
...
    AccountExport:
      type: object
      required: [exportedAt, account, tasks, links, tags, achievements]
      properties:
        exportedAt: { type: string, format: date-time }
        account: { $ref: "#/components/schemas/User" }
        tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
        links: { type: array, items: { $ref: "#/components/schemas/TaskLink" } }
        tags: { type: array, items: { $ref: "#/components/schemas/Tag" } }
        achievements:
          type: array
          items:
            type: object
            required: [code, unlockedAt]
            properties:
              code: { type: string }
              unlockedAt: { type: string, format: date-time }
...
    Tag:
      type: object
      required: [id, name, taskCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 24 }
        taskCount: { type: integer }
    LinkType:
      type: string
      enum: [RELATED, CONNECTED_TO, BLOCKS, DEPENDS_ON]
...
    TaskLink:
      type: object
      required: [id, sourceTaskId, targetTaskId, type, createdAt]
      properties:
        id: { type: string, format: uuid }
        sourceTaskId: { type: string, format: uuid }
        targetTaskId: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/LinkType" }
        createdAt: { type: string, format: date-time }
...
    Task:
      type: object
      description: |
        A regular task or a subtask; parentTaskId tells them apart. quadrant is
        non-null exactly when parentTaskId is null: a subtask has none of its
        own and inherits the parent's. The four completion fields are set
        together, exactly when status is COMPLETED, and are frozen at that
        moment so a later change to the XP configuration never rewrites
        history.
      required: [id, userId, title, description, position, color, deadlineHasTime, status, createdAt, updatedAt, tags]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        parentTaskId: { type: string, format: uuid, nullable: true }
        title: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 2000 }
        quadrant: { allOf: [{ $ref: "#/components/schemas/Quadrant" }], nullable: true }
        position: { type: integer, format: int32 }
        color: { $ref: "#/components/schemas/TaskColor" }
        deadlineAt: { type: string, format: date-time, nullable: true }
        deadlineHasTime:
          type: boolean
          description: False means a date-only deadline, rendered without a time.
        status: { $ref: "#/components/schemas/TaskStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        xpAwarded: { type: integer, format: int32, nullable: true }
        quadrantAtCompletion: { allOf: [{ $ref: "#/components/schemas/Quadrant" }], nullable: true }
        completedVia: { allOf: [{ $ref: "#/components/schemas/CompletionSource" }], nullable: true }
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Always absent on the board, which never shows a deleted task. It
            is here for the export, which includes them — a list that carried
            them without saying which would be worse than leaving them out.
        tags: { type: array, items: { type: string } }
    Plan:
      type: string
      enum: [FREE, PRO, SELF_HOSTED]
    User:
      type: object
      required:
        [id, username, language, timezone, showInLeaderboard, plan, planExpired, createdAt, stats]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email:
          type: string
          format: email
          nullable: true
          description: Absent until the user adds one.
        avatarUrl:
          type: string
          nullable: true
          description: |
            Null means the client renders an identicon from the username. The
            server does not generate one.
        language: { $ref: "#/components/schemas/Language" }
        timezone: { type: string }
        showInLeaderboard: { type: boolean }
        plan: { $ref: "#/components/schemas/Plan" }
        planExpired:
          type: boolean
          description: A trial that has run out. Every section but pricing is closed.
        createdAt: { type: string, format: date-time }
        stats: { $ref: "#/components/schemas/UserStats" }
    UserStats:
      type: object
      description: |
        A recomputable aggregate of the XP ledger and the activity log, read on
        every request rather than summed from the ledger each time.
      required:
        [lifetimeXp, level, currentStreak, longestStreak, tasksCreated, tasksCompleted,
         subtasksCompleted, linksCreated, achievementsUnlocked]
      properties:
        lifetimeXp: { type: integer, format: int64 }
        level: { type: integer, format: int32, minimum: 1 }
        currentStreak: { type: integer, format: int32 }
        longestStreak: { type: integer, format: int32 }
        tasksCreated: { type: integer, format: int32 }
        tasksCompleted: { type: integer, format: int32 }
        subtasksCompleted: { type: integer, format: int32 }
        linksCreated: { type: integer, format: int32 }
        achievementsUnlocked: { type: integer, format: int32 }
security:
  - accessCookie: []
    csrfToken: []